requirements.txt
pandas~=2.1.0
openai~=0.28.0
python-dotenv~=1.0.0
tqdm~=4.66.1